fix compile warning in demo
```
In file included from demo/keyboard.c:5:
demo/keyboard.c: In function ‘draw_keyboard’:
demo/../termbox2.h:252:33: warning: unsigned conversion from ‘int’ to ‘uint16_t’ {aka ‘short unsigned int’} changes value from ‘16777222’ to ‘6’ [-Woverflow]
  252 | #define TB_MAGENTA              0x0006
      |                                 ^
demo/keyboard.c:512:22: note: in expansion of macro ‘TB_MAGENTA’
  512 |     printf_tb(33, 1, TB_MAGENTA | TB_BOLD, TB_DEFAULT, "Keyboard demo!");
```

+++ demo/keyboard.c
@@ -509,7 +509,7 @@ void draw_keyboard()
     draw_key(K_K_0, TB_WHITE, TB_BLUE);
     draw_key(K_K_PERIOD, TB_WHITE, TB_BLUE);
 
-    printf_tb(33, 1, TB_MAGENTA | TB_BOLD, TB_DEFAULT, "Keyboard demo!");
+    printf_tb(33, 1, (uint16_t)(TB_MAGENTA | TB_BOLD), TB_DEFAULT, "Keyboard demo!");
     printf_tb(21, 2, TB_MAGENTA, TB_DEFAULT, "(press CTRL+X and then CTRL+Q to exit)");
     printf_tb(15, 3, TB_MAGENTA, TB_DEFAULT, "(press CTRL+X and then CTRL+C to change input mode)");
